sandbox/solid_python_practice/pyproject.toml
[tool.poetry]
name = "solid-python-practice"
version = "0.1.0"
description = ""
authors = ["Masahiro Wada <argon.argon.argon@gmail.com>"]
readme = "README.md"
packages = [{include = "solid_python_practice"}]

[tool.poetry.dependencies]
python = "^3.10"
solidpython = "^1.1.3"
viewscad = "^0.2.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
